backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "kitfjn"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "uuid (>=1.30,<2.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pgcli (>=4.3.0,<5.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
